разделяет на модули
